# swagger/spec.json
{"openapi":"3.0.0","info":{"title":"Pets Services API","version":"1.0.0"},"paths":{"/absences":{"post":{"tags":["absence"],"security":[{"bearerAuth":[]}],"description":"User add a new absence","requestBody":{"$ref":"#/components/requestBodies/AddAbsence"},"responses":{"201":{"$ref":"#/components/responses/AbsenceAdded"},"400":{"description":"Bad request"},"401":{"$ref":"#/components/responses/MissingJWT"},"500":{"description":"Server error"}}},"get":{"tags":["absence"],"security":[{"bearerAuth":[]}],"description":"Get absences of a user","responses":{"200":{"$ref":"#/components/responses/AbsencesReturned"},"400":{"description":"Bad request"},"401":{"$ref":"#/components/responses/MissingJWT"},"500":{"description":"Server error"}}}},"/absences/{id}":{"delete":{"tags":["absence"],"security":[{"bearerAuth":[]}],"description":"User delete one of his absences","parameters":[{"name":"id","description":"Absence id","in":"path","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"$ref":"#/components/responses/AbsenceDeleted"},"400":{"description":"Bad request"},"401":{"$ref":"#/components/responses/MissingJWT"},"500":{"description":"Server error"}}}},"/animals":{"get":{"tags":["animal"],"security":[{"bearerAuth":[]}],"description":"Get animals of a user","responses":{"200":{"$ref":"#/components/responses/AnimalsReturned"},"400":{"description":"Bad request"},"500":{"description":"Server error"}}},"post":{"tags":["animal"],"security":[{"bearerAuth":[]}],"description":"User add a new animal","requestBody":{"$ref":"#/components/requestBodies/AddAnimal"},"responses":{"201":{"$ref":"#/components/responses/AnimalAdded"},"400":{"description":"Bad request"},"401":{"$ref":"#/components/responses/MissingJWT"},"500":{"description":"Server error"}}}},"/animals/{id}":{"delete":{"tags":["animal"],"description":"Delete animal","security":[{"bearerAuth":[]}],"parameters":[{"name":"id","description":"Animal id","in":"path","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"$ref":"#/components/responses/AnimalDeleted"},"400":{"description":"Bad request"},"401":{"$ref":"#/components/responses/MissingJWT"},"404":{"$ref":"#/components/responses/AnimalNotDeleted"},"500":{"description":"Server error"}}},"put":{"tags":["animal"],"security":[{"bearerAuth":[]}],"parameters":[{"name":"id","description":"Animal id","in":"path","required":true,"schema":{"type":"integer"}}],"description":"User update one of his animals","requestBody":{"$ref":"#/components/requestBodies/UpdateAnimal"},"responses":{"201":{"$ref":"#/components/responses/AnimalUpdated"},"400":{"description":"Bad request"},"401":{"$ref":"#/components/responses/MissingJWT"},"500":{"description":"Server error"}}}},"/animalTypes":{"get":{"tags":["animal type"],"security":[{"bearerAuth":[]}],"description":"Get all the animal types","responses":{"200":{"$ref":"#/components/responses/AnimalTypesAreFound"},"400":{"description":"Bad request"},"401":{"$ref":"#/components/responses/MissingJWT"},"500":{"description":"Server error"}}}},"/animalTypes/supplier":{"get":{"tags":["animal type"],"security":[{"bearerAuth":[]}],"description":"Get all the animal types of a specific supplier","responses":{"200":{"$ref":"#/components/responses/SupplierAnimalTypesAreFound"},"400":{"description":"Bad request"},"401":{"$ref":"#/components/responses/MissingJWT"},"500":{"description":"Server error"}}},"post":{"tags":["animal type"],"security":[{"bearerAuth":[]}],"description":"Add a supplier animal type","requestBody":{"$ref":"#/components/requestBodies/UserAnimaltype"},"responses":{"201":{"$ref":"#/components/responses/UserAnimaltypeAdded"},"400":{"description":"Bad request"},"401":{"$ref":"#/components/responses/MissingJWT"},"500":{"description":"Server error"}}}},"/animalTypes/supplier/{id}":{"delete":{"tags":["animal type"],"security":[{"bearerAuth":[]}],"description":"Delete a supplier animal type accepted","parameters":[{"name":"id","description":"animal type id","in":"path","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"$ref":"#/components/responses/UserAnimaltypeDeleted"},"400":{"description":"Bad request"},"401":{"$ref":"#/components/responses/MissingJWT"},"403":{"$ref":"#/components/responses/UserAnimaltypeNotDeleted"},"500":{"description":"Server error"}}}},"/serviceshours":{"post":{"tags":["service hours"],"security":[{"bearerAuth":[]}],"description":"A service hours is created bu the customer","requestBody":{"$ref":"#/components/requestBodies/AddServicesHour"},"responses":{"201":{"$ref":"#/components/responses/ServiceHoursAdded"},"400":{"description":"Bad request"},"401":{"$ref":"#/components/responses/MissingJWT"},"409":{"$ref":"#/components/responses/ServiceHoursNotAvailable"},"500":{"description":"Server error"}}},"get":{"tags":["service hours"],"description":"Get services hours which have specific data","security":[{"bearerAuth":[]}],"responses":{"200":{"$ref":"#/components/responses/ServicesHoursAreFound"},"400":{"description":"Bad request"},"401":{"$ref":"#/components/responses/MissingJWT"},"403":{"$ref":"#/components/responses/mustBeAdministrator"},"500":{"description":"Server error"}}}},"/serviceshours/{id}":{"patch":{"tags":["service hours"],"security":[{"bearerAuth":[]}],"description":"A service hours is updated by the supplier","parameters":[{"name":"id","description":"Service hours id","in":"path","required":true,"schema":{"type":"integer"}}],"requestBody":{"$ref":"#/components/requestBodies/UpdateServiceHours"},"responses":{"201":{"$ref":"#/components/responses/ServiceHoursUpdated"},"400":{"description":"Bad request"},"401":{"$ref":"#/components/responses/MissingJWT"},"500":{"description":"Server error"}}},"delete":{"tags":["service hours"],"security":[{"bearerAuth":[]}],"description":"Admin delete one of services hours","parameters":[{"name":"id","description":"Service hours id","in":"path","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"$ref":"#/components/responses/ServiceHoursDeleted"},"400":{"description":"Bad request"},"401":{"$ref":"#/components/responses/MissingJWT"},"403":{"$ref":"#/components/responses/mustBeAdministrator"},"500":{"description":"Server error"}}}},"/serviceshours/customer":{"get":{"tags":["service hours"],"security":[{"bearerAuth":[]}],"description":"Get services hours as a customer","responses":{"200":{"$ref":"#/components/responses/CustomerServicesHoursReturned"},"400":{"description":"Bad request"},"401":{"$ref":"#/components/responses/MissingJWT"},"500":{"description":"Server error"}}}},"/serviceshours/supplier":{"get":{"tags":["service hours"],"security":[{"bearerAuth":[]}],"description":"Get services hours as a supplier","responses":{"200":{"$ref":"#/components/responses/SupplierServicesHoursReturned"},"400":{"description":"Bad request"},"401":{"$ref":"#/components/responses/MissingJWT"},"500":{"description":"Server error"}}}},"/users":{"post":{"tags":["user"],"description":"Create new user and return JWT for user identification","requestBody":{"$ref":"#/components/requestBodies/AddUser"},"responses":{"201":{"$ref":"#/components/responses/UserRegistered"},"400":{"$ref":"#/components/responses/UserNotRegistered"},"409":{"$ref":"#/components/responses/UserAlreadyExist"},"500":{"description":"Server error"}}},"get":{"tags":["user"],"description":"Get users who have specific data","security":[{"bearerAuth":[]}],"responses":{"200":{"$ref":"#/components/responses/UsersAreFound"},"400":{"description":"Bad request"},"401":{"$ref":"#/components/responses/MissingJWT"},"403":{"$ref":"#/components/responses/mustBeAdministrator"},"500":{"description":"Server error"}}}},"/users/actions/login":{"post":{"tags":["user"],"description":"User login and return JWT for user identification","requestBody":{"description":"Email and password for user connection","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Login"}}}},"responses":{"201":{"$ref":"#/components/responses/LoginAccepted"},"401":{"$ref":"#/components/responses/LoginRejected"},"500":{"description":"Server error"}}}},"/users/{id}":{"put":{"tags":["user"],"security":[{"bearerAuth":[]}],"description":"Update data of a specific user","parameters":[{"name":"id","description":"User id","in":"path","required":true,"schema":{"type":"integer"}}],"requestBody":{"$ref":"#/components/requestBodies/UpdateUser"},"responses":{"200":{"$ref":"#/components/responses/UserUpdated"},"400":{"description":"Bad request"},"401":{"$ref":"#/components/responses/MissingJWT"},"403":{"$ref":"#/components/responses/UserDoesNotHaveAcces"},"500":{"description":"Server error"}}},"patch":{"tags":["user"],"security":[{"bearerAuth":[]}],"description":"Update data of a specific user by admin","parameters":[{"name":"id","description":"User id","in":"path","required":true,"schema":{"type":"integer"}}],"requestBody":{"$ref":"#/components/requestBodies/UpdateUserByAdmin"},"responses":{"200":{"$ref":"#/components/responses/UserUpdatedByAdmin"},"400":{"description":"Bad request"},"401":{"$ref":"#/components/responses/MissingJWT"},"403":{"$ref":"#/components/responses/mustBeAdministrator"},"500":{"description":"Server error"}}},"get":{"tags":["user"],"security":[{"bearerAuth":[]}],"description":"Get data of a specific user","parameters":[{"name":"id","description":"User id","in":"path","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"$ref":"#/components/responses/UserIsFound"},"400":{"description":"Bad request"},"401":{"$ref":"#/components/responses/MissingJWT"},"403":{"$ref":"#/components/responses/UserDoesNotHaveAcces"},"500":{"description":"Server error"}}},"delete":{"tags":["user"],"description":"Delete user","security":[{"bearerAuth":[]}],"parameters":[{"name":"id","description":"User id","in":"path","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"$ref":"#/components/responses/UserDeleted"},"400":{"description":"Bad request"},"401":{"$ref":"#/components/responses/MissingJWT"},"403":{"$ref":"#/components/responses/mustBeAdministrator"},"404":{"$ref":"#/components/responses/UserNotDeleted"},"500":{"description":"Server error"}}}},"/users/actions/addbyadmin":{"post":{"tags":["user"],"description":"Create new user as an administrator","security":[{"bearerAuth":[]}],"requestBody":{"$ref":"#/components/requestBodies/AddUserByAdmin"},"responses":{"201":{"$ref":"#/components/responses/UserRegisteredByAdmin"},"400":{"description":"Bad request"},"401":{"$ref":"#/components/responses/MissingJWT"},"403":{"$ref":"#/components/responses/mustBeAdministrator"},"409":{"$ref":"#/components/responses/UserAlreadyExist"},"500":{"description":"Server error"}}}}},"components":{"responses":{"AbsenceAdded":{"description":"User Absence added"},"AbsenceDeleted":{"description":"User Absence deleted"},"AbsencesReturned":{"description":"The absences are returned"},"AnimalsReturned":{"description":"The animals are returned"},"AnimalAdded":{"description":"User Animal added"},"AnimalUpdated":{"description":"User Animal updated"},"AnimalDeleted":{"description":"The animal is deleted"},"AnimalNotDeleted":{"description":"The animal is not found"},"AnimalTypesAreFound":{"description":"The animal types are returned"},"AnimalTypesAreNotFound":{"description":"The animal types are not obtained"},"UserAnimaltypeAdded":{"description":"The supplier animal type is added"},"SupplierAnimalTypesAreFound":{"description":"The supplier animal types are returned"},"UserAnimaltypeDeleted":{"description":"The supplier animal type is deleted"},"UserAnimaltypeNotDeleted":{"description":"The supplier animal type id is not found"},"ServiceHoursAdded":{"description":"Service Hour added"},"ServiceHoursNotAvailable":{"description":"Service Hour not available"},"ServiceHoursUpdated":{"description":"Service Hour updated"},"CustomerServicesHoursReturned":{"description":"The customer services hours are returned"},"SupplierServicesHoursReturned":{"description":"The supplier services hours are returned"},"ServiceHoursDeleted":{"description":"Service Hours deleted"},"ServicesHoursAreFound":{"description":"The services hours are returned"},"UserRegistered":{"description":"The user is registered and received his jwt"},"UserAlreadyExist":{"description":"The user already exists"},"UserNotRegistered":{"description":"Invalid inputs"},"LoginAccepted":{"description":"The user is connected and recieved jwt"},"LoginRejected":{"description":"The user is not connected because of bad email/password"},"UserIsFound":{"description":"The user is returned"},"UserUpdated":{"description":"The user is updated"},"UserDoesNotHaveAcces":{"description":"The user does not have access"},"UserUpdatedByAdmin":{"description":"The user is updated by admin"},"UsersAreFound":{"description":"The users are returned"},"UserAreNotFound":{"description":"Thee users are not obtained"},"UserDeleted":{"description":"The user is deleted"},"UserNotDeleted":{"description":"The user is not found"},"UserRegisteredByAdmin":{"description":"The user is registered"},"UserNotRegisteredByAdmin":{"description":"Invalid inputs"},"mustBeAdministrator":{"description":"The requested action can only be performed by an administrator"},"MissingJWT":{"description":"JWT is not present"}},"requestBodies":{"AddAbsence":{"content":{"application/json":{"schema":{"type":"object","properties":{"date":{"type":"string"}},"required":["date"]}}}},"AddAnimal":{"content":{"application/json":{"schema":{"type":"object","properties":{"breed":{"type":"string"},"review":{"type":"string"},"weight":{"type":"string"},"name":{"type":"string"},"animalTypeId":{"type":"integer"}},"required":["breed","weight","name","animalTypeId"]}}}},"UpdateAnimal":{"content":{"application/json":{"schema":{"type":"object","properties":{"breed":{"type":"string"},"review":{"type":"string"},"weight":{"type":"string"},"name":{"type":"string"},"animalTypeId":{"type":"integer"}}}}}},"UserAnimaltype":{"content":{"application/json":{"schema":{"type":"object","properties":{"animalTypeId":{"type":"integer"}},"required":["animalTypeId"]}}}},"AddServicesHour":{"content":{"application/json":{"schema":{"type":"object","properties":{"startDateTime":{"type":"string"},"endDateTime":{"type":"string"},"type":{"type":"string"},"description":{"type":"string"},"emailSupplier":{"type":"string"},"animalId":{"type":"integer"}},"required":["startDateTime","endDateTime","type","emailSupplier","animalId"]}}}},"UpdateServiceHours":{"content":{"application/json":{"schema":{"type":"object","properties":{"description":{"type":"string"},"isAcceptedRequest":{"type":"boolean"}},"required":["isAcceptedRequest"]}}}},"AddUser":{"content":{"application/json":{"schema":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"firstname":{"type":"string"},"lastname":{"type":"string"},"phone":{"type":"string"},"locality":{"type":"string"},"postalCode":{"type":"integer"},"streetNumber":{"type":"string"},"streetName":{"type":"string"},"country":{"type":"string"},"customer":{"type":"object","properties":{"searchWalker":{"type":"boolean"},"searchHost":{"type":"boolean"}},"required":["searchWalker","searchHost"]},"supplier":{"type":"object","properties":{"isHost":{"type":"boolean"},"isAnimalWalker":{"type":"boolean"}},"required":["isHost","isAnimalWalker"]}},"required":["email","password","firstname","lastname","phone","locality","postalCode","streetNumber","streetName","country"]}}}},"UpdateUser":{"content":{"application/json":{"schema":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"firstname":{"type":"string"},"lastname":{"type":"string"},"phone":{"type":"string"},"locality":{"type":"string"},"postalCode":{"type":"integer"},"streetNumber":{"type":"string"},"streetName":{"type":"string"},"country":{"type":"string"},"picture":{"type":"string"},"customer":{"type":"object","properties":{"searchWalker":{"type":"boolean"},"searchHost":{"type":"boolean"},"commune":{"type":"string"}},"required":["searchWalker","searchHost"]},"supplier":{"type":"object","properties":{"isHost":{"type":"boolean"},"isAnimalWalker":{"type":"boolean"},"slogan":{"type":"string"},"commune":{"type":"string"},"weightMax":{"type":"integer"}},"required":["isHost","isAnimalWalker"]}},"required":["email","password","firstname","lastname","phone","locality","postalCode","streetNumber","streetName","country"]}}}},"UpdateUserByAdmin":{"content":{"application/json":{"schema":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"firstname":{"type":"string"},"lastname":{"type":"string"},"phone":{"type":"string"},"locality":{"type":"string"},"postalCode":{"type":"integer"},"streetNumber":{"type":"string"},"streetName":{"type":"string"},"country":{"type":"string"},"customer":{"type":"object","properties":{"searchWalker":{"type":"boolean"},"searchHost":{"type":"boolean"}},"required":["searchWalker","searchHost"]},"supplier":{"type":"object","properties":{"isHost":{"type":"boolean"},"isAnimalWalker":{"type":"boolean"}},"required":["isHost","isAnimalWalker"]}},"required":["email","password","firstname","lastname","phone","locality","postalCode","streetNumber","streetName","country"]}}}},"AddUserByAdmin":{"content":{"application/json":{"schema":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"firstname":{"type":"string"},"lastname":{"type":"string"},"phone":{"type":"string"},"locality":{"type":"string"},"postalCode":{"type":"integer"},"streetNumber":{"type":"string"},"streetName":{"type":"string"},"country":{"type":"string"},"customer":{"type":"object","properties":{"searchWalker":{"type":"boolean"},"searchHost":{"type":"boolean"}},"required":["searchWalker","searchHost"]},"supplier":{"type":"object","properties":{"isHost":{"type":"boolean"},"isAnimalWalker":{"type":"boolean"}},"required":["isHost","isAnimalWalker"]}},"required":["email","password","firstname","lastname","phone","locality","postalCode","streetNumber","streetName","country"]}}}}},"schemas":{"Login":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}},"tags":[]}